moving the off-axis sas implementation to a separate file + additional changes to fix usage with cupy
