+++ PtyLab/Operators/Operators.py
@@ -13,11 +13,12 @@ except ImportError:
     pass
     # print("Cupy not available, will not be able to run GPU based computation")
 import numpy as np
-from scipy.signal import convolve2d
-from scipy.signal.windows import gaussian
 
 from PtyLab import Params, Reconstruction
 from PtyLab.Operators._propagation_kernels import __make_quad_phase
+from PtyLab.Operators.off_axis_sas import (
+    __make_transferfunction_off_axis_sas, propagate_off_axis_sas)
+from PtyLab.Operators.propagator_utils import complexexp
 from PtyLab.utils.gpuUtils import getArrayModule, isGpuArray
 from PtyLab.utils.utils import circ, fft2c, ifft2c
 
@@ -550,277 +551,6 @@ def propagate_polychromeASP_inv(fields, params, reconstruction, z=None):
     """
     return propagate_polychromeASP(fields, params, reconstruction, inverse=True, z=z)
 
-
-def _iterate_6d_fields(fields):
-    """Efficiently iterating via a generator for the 6D field array
-    (nlambda, nosm, npsm, nslice, Np, Np).
-
-    Parameters
-    ----------
-    fields : np.ndarray
-
-
-    Yields
-    ------
-    tuple
-        returns a tuple of indices over the array
-    """
-    for i_nlambda in range(fields.shape[0]):
-        for j_nosm in range(fields.shape[1]):
-            for k_npsm in range(fields.shape[2]):
-                for l_nslice in range(fields.shape[3]):
-                    yield (i_nlambda, j_nosm, k_npsm, l_nslice)
-
-
-def propagate_off_axis_sas(
-    fields: np.ndarray,
-    params: Params,
-    reconstruction: Reconstruction,
-    z: float = None,
-):
-    """
-    Scalable Off-axis Angular Spectrum (SOAS) Propagation method that assumes that the source and
-    destination planes are coplanar, but off-axis.
-
-    Parameters
-    ----------
-    fields: np.ndarray
-        Field to propagate
-    params: Params
-        Instance of the Params class
-    reconstruction: Reconstruction
-        Instance of the Reconstruction class.
-    z: float
-        Propagation distance
-
-    Returns
-    -------
-    reconstruction.esw, propagated field:
-        Exit surface wave and the propagated field
-    """
-
-    xp = getArrayModule(fields)
-
-    # pad the original field (last 2 dimensions) with zeros to be twice it's size
-    # no padding in the first 4 dimensions
-    rows, cols = fields.shape[-2:]
-    pad_factor = 2
-    pad_width = (
-        (0, 0),
-        (0, 0),
-        (0, 0),
-        (0, 0),
-        (rows // pad_factor, rows // pad_factor),
-        (cols // pad_factor, cols // pad_factor),
-    )
-    fields_padded = xp.pad(fields, pad_width, "constant")
-
-    # reconstruction parameters
-    Np = reconstruction.Np
-    dxp = reconstruction.dxp
-    wavelength = reconstruction.wavelength
-    z = reconstruction.zo if z is None else z
-
-    # quadratic phase Q2 (currently zo, but this can be z2 and z1 separated)
-    quad_phase = __make_quad_phase(
-        z, wavelength, Np * pad_factor, dxp, params.gpuSwitch
-    )
-
-    # precompensated transfer function
-    H_precomp = __make_transferfunction_off_axis_sas(
-        params, reconstruction, pad_factor, z
-    )
-
-    # field propagation
-    psi_precomp = ifft2c(H_precomp * fft2c(fields_padded))
-    prop_fields = fft2c(quad_phase * psi_precomp)
-
-    # crop the field by a factor of 2 (as it was originally padded by 2)
-    rows_padded, cols_padded = prop_fields.shape[-2:]
-    start_h, start_w = (
-        (rows_padded - rows) // 2,
-        (cols_padded - cols) // 2,
-    )
-    slicey, slicex = (
-        slice(start_h, start_h + rows),
-        slice(start_w, start_w + cols),
-    )
-    prop_fields = prop_fields[..., slicey, slicex]
-
-    return reconstruction.esw, prop_fields
-
-
-@lru_cache(cache_size)
-def __make_transferfunction_off_axis_sas(
-    params: Params,
-    reconstruction: Reconstruction,
-    pad_factor: int,
-    zo: float = None,
-):
-    """
-    Allows for a 6-dimensional (nlambda, nosm, npsm, nslice, Np, Np) array when computing the transfer function
-    for a scalable off-axis angular spectrum propagator.
-
-    Parameters
-    ----------
-    params: Params
-        Instance of the Params class
-    reconstruction: Reconstruction
-        Instance of the Reconstruction class.
-    z: float
-        Propagation distance
-
-    Returns
-    -------
-    np.ndarray or cp.ndarray
-        The calculated transfer function with shape (nlambda, nosm, npsm, nslice, Np, Np).
-    """
-
-    fftshiftSwitch = params.fftshiftSwitch
-    Np = reconstruction.Np  # Pixel size along each dimension.
-    wavelength = reconstruction.wavelength  # Wavelength used in the scanning probe.
-    theta = reconstruction.theta
-    Lp = reconstruction.Lp  # length of the sample.
-    nosm = reconstruction.nosm  # no. of spatial modes for the object.
-    npsm = reconstruction.npsm  # no. of spatial modes for the probe.
-    nlambda = reconstruction.nlambda  # no. of wavelengths for multi-wavelength.
-    nslice = reconstruction.nslice  # no. of slices for multi-slice operation
-    on_gpu = params.gpuSwitch  # switch to use GPU acceleration.
-
-    # distance from the center of the sample to the probe.
-    zo = reconstruction.zo if zo is None else zo
-
-    # ensuring some checks
-    if fftshiftSwitch:
-        raise ValueError("ASP propagatorType works only with fftshiftSwitch = False!")
-
-    if nlambda > 1:
-        raise ValueError("Currently for multi-wavelength, off-axis SAS does not work")
-
-    if nslice > 1:
-        raise ValueError(
-            " Currently off-axis SAS not valid for multi-slice ptychography"
-        )
-
-    # Array shape (nlambda, nosm, npsm, nslice, Np, Np)
-    transfer_function = np.zeros(
-        (nlambda, nosm, npsm, nslice, Np * pad_factor, Np * pad_factor),
-        dtype="complex64",
-    )
-    for inds in _iterate_6d_fields(transfer_function):
-        transfer_function[*inds] = __off_axis_sas_transfer_function(
-            wavelength, Lp, pad_factor * Np, theta, zo, on_gpu
-        )
-
-    return (
-        cp.array(transfer_function) if on_gpu and cp is not None else transfer_function
-    )
-
-
-@lru_cache(cache_size)
-def __off_axis_sas_transfer_function(wavelength, Lp, Np, theta, zo, on_gpu):
-    """Precompensation transfer function for scalable off-axis transfer function.
-
-    Parameters
-    ----------
-    wavelength : float
-        wavelength
-    Lp : float
-        Physical size
-    Np : float
-        _description_
-    theta : tuple / scalar
-        Theta (angle in degrees) in the x-y plane.
-    zo : float
-        propagation distance
-    on_gpu : bool
-        checks if the array is on GPU or not.
-
-    Returns
-    -------
-    np.ndarray
-        precompensated transfer function `H_precomp`
-    """
-    # TODO: Test with and without this caching mechanism.
-
-    # cp/np array
-    xp = cp if on_gpu else np
-
-    # Fourier grid
-    df = 1 / Lp
-    f = xp.arange(-Np / 2, Np / 2) * df
-    Fx, Fy = xp.meshgrid(f, f)
-
-    # off-axis sines and tangents (theta in degrees)
-    thetax, thetay = theta
-    # TODO: theta is defined with regards to aPIE, however, it is usually a scalar
-    # float, perhaps we should specify it as a tuple (thetax,thetay) - or something
-    # better? or simply create a new attribute?
-    sx = xp.sin(xp.radians(thetax))
-    sy = xp.sin(xp.radians(thetay))
-    tx = xp.tan(xp.radians(thetax))
-    ty = xp.tan(xp.radians(thetay))
-
-    # transfer function
-    # eq. 12 includes chi parameter under square root
-    chi = (
-        1 / wavelength**2
-        - (Fx + (sx / wavelength)) ** 2
-        - (Fy + (sy / wavelength)) ** 2
-    )
-    sqrt_chi = np.sqrt(np.maximum(0, chi))
-
-    # creating a bandpass filter
-
-    def _create_bandpass_filter(smooth_filter=True, eps=1e-10):
-        # zo = z1 in the first line and zo = z2 in the second line
-        Omegax = zo * (tx - (Fx + sx / wavelength) / (sqrt_chi + eps))
-        Omegax += wavelength * zo * Fx
-
-        # zo = z1 in the first line and zo = z2 in the second line
-        Omegay = zo * (ty - (Fy + sy / wavelength) / (sqrt_chi + eps))
-        Omegay += wavelength * zo * Fy
-
-        sampling_rate = 2
-        criteria_x = df <= xp.abs(1 / (sampling_rate * Omegax + eps))
-        criteria_y = df <= xp.abs(1 / (sampling_rate * Omegay + eps))
-
-        # Fourier Bandpass filter (W is a mask below)
-        W_mask = xp.logical_and(criteria_x, criteria_y)
-
-        # Gaussian kernel
-        gauss_1d = gaussian(8, 2)
-        gauss_1d /= xp.sum(gauss_1d)
-        kernel_gauss = xp.outer(gauss_1d, gauss_1d)
-
-        # smooth the bandpass filter corners with a Gaussian kernel
-        bandpass_filter = (
-            convolve2d(W_mask, kernel_gauss, mode="same") if smooth_filter else W_mask
-        )
-
-        return bandpass_filter
-
-    # Pre-compensation transfer function
-
-    # implements the angular spectrum transfer function (see eq. 23, part of the precompensation factor)
-    # zo is z1 in the document.
-    H_AS = complexexp(2 * np.pi * zo * sqrt_chi)
-
-    # Fresnel transfer function
-    H_Fr = complexexp(
-        -np.pi * zo / wavelength * ((wavelength * Fx) ** 2 + (wavelength * Fy) ** 2)
-    )
-
-    # off-axis consideration of the transfer function
-    H_offaxis = complexexp(2 * np.pi * zo * (tx * Fx + ty * Fy))
-
-    # precompensation with bandpass filter
-    bandpass_filter = _create_bandpass_filter(smooth_filter=True, eps=1e-10)
-    H_precomp = H_AS * xp.conj(H_Fr) * H_offaxis * bandpass_filter
-
-    return H_precomp
-
-
 def detector2object(fields, params: Params, reconstruction: Reconstruction):
     """
     Implements detector2object.m. Returns a propagated version of the field.
@@ -889,25 +619,6 @@ def aspw(u, z, wavelength, L, bandlimit=True, is_FT=True):
         U = fft2c(u)
     u_prop = ifft2c(U * phase_exp)
     return u_prop, phase_exp
-
-
-def complexexp(angle):
-    """
-    Faster way of implementing np.exp(1j*something_unitary)
-
-    Parameters
-    ----------
-    angle: np.ndarray
-        Angle of the exponent
-
-    Returns
-    -------
-
-    cos(angle) + 1j * sin(angle)
-
-    """
-    xp = getArrayModule(angle)
-    return xp.cos(angle) + 1j * xp.sin(angle)
 
 
 def scaledASP(u, z, wavelength, dx, dq, bandlimit=True, exactSolution=False):
